Fixes how profdata_coverage_dir is created.
The previous implementation of this method was assuming that
the first product always has name, which doesn't always happen.

The documentation of PBXFileReference says that the 'name'
attribute is "often not present.". For more info check
https://github.com/CocoaPods/Xcodeproj/blob/master/lib/xcodeproj/project/object/file_reference.rb#L13

If the name attribute is not present the profdata_coverage_dir
created was the first child of the build directory (which by
default is the derived data directory). This is wrong.

Also the build directory was being concatenated using plain
string manipulation which led to path of the form:
"~/Library/Developer/Xcode/DerivedData//**/PRODUCT_NAME".
This is why File.join is used to join the build directory path
with '/**/PRODUCT_NAME' to comtemplate the case of build directory
already ending with '/'.