Fix bad UTF-8 encoding issues
Before this commit, the html report was displaying this:
//  Copyright (c) 2013-2016 C��dric Luthi. All rights reserved.

After this commit, the html report is displaying this:
//  Copyright (c) 2013-2016 Cédric Luthi. All rights reserved.

I added a 🌟 (Glowing Star, U+1F31F) in the fixtures to make sure it’s property encoded.

The problem was that valid UTF-8 was interpreted as "binary" and re-encoded as UTF-8, producing invalid output.