Use `llvm-cov export` to detect source files per binary
When binary contains single source file, output of `llvm-cov show` does not contain the file name. That leads to `source_file_pathname` to contain actual source code (first line of it) instead of a path.

Previously, result of `find_source_files` method was used to understand if a binary is a single source:
```
source_files = find_source_files || []
…
# If a single source file is used, the resulting output does not contain the file name.
coverage_file.source_file_pathname = source_files.first if source_files.count == 1
```
Though that approach won’t work for projects which contains several binaries (and then `source_files` will be a mix of files from different binaries) or when you don’t provide list of source file in configuration (it’s what `find_source_files` returns). So we either need to provide the list of files per binary or somehow get the info from profdata file.

`llvm-cov export` tool exports the coverage data in JSON format and it doesn’t depend on whether there is a single source file or more in a binary. So I propose to use it and parse JSON output to get a list of source files prior to actually running `llvm-cov show` command.

That makes optimizations on lines 134-138 obsolete.